Store hashes of avatar photos to avoid spurious updates
When the bridge is restarted, Google Chat sends different
URLs for the avatar images for every participant in every room.
This happens even if the images have not changed, but sending
'avatar change' events into the Matrix rooms can be costly.

This patch computes a SHA256 hash of the image to see if the
image has actually changed; if not, the new URL is stored but no
Matrix event is sent.

This includes a database schema upgrade to v4.