Get the bridge working again
So there is a /u/0/api endpoint that I discovered yesterday that has literally
the exact same api we were using before. So I reverted all of the batchexecute
stuff, restored the protobuf file and everything seems to work. We still need
to do the cookie auth which sucks, and make the request to /u/0/mole/world to
get some magic values, but everything seems to be working fine in my limited
testing so far.